pyproject.toml
[tool.poetry]
name = "brian"
version = "0.1.0"
description = ""
authors = ["Sean Lim <seanlimdev856@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.1"
marshmallow = "^3.2"
face-recognition = "^1.2"
marshmallow-sqlalchemy = "^0.19.0"
psycopg2 = "^2.8"
python-dotenv = "^0.10.3"
bcrypt = "^3.1.7"
flask-cors = "^3.0.8"
alembic = "^1.3"
loguru = "^0.4.0"
flask-jwt-extended = "^3.24"
Flask-User = "^1.0"
Flask-SQLAlchemy = "^2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
